Rename InputFormView to RenameForm
